+++ frontend/index.html
@@ -1,18 +1,17 @@
 <!DOCTYPE html>
-<html>
-
-<head>
-  <meta charset="UTF-8">
-  <!-- https://developer.mozilla.org/en-US/docs/Web/HTTP/CSP -->
-  <meta http-equiv="Content-Security-Policy" content="default-src 'self'; script-src 'self'">
-  <title>Hello World!</title>
-</head>
-
-<body>
-  <h1>Hello World!</h1>
-  We are using Node.js <span id="node-version"></span>,
-  Chromium <span id="chrome-version"></span>,
-  and Electron <span id="electron-version"></span>.
-</body>
-
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta
+      http-equiv="Content-Security-Policy"
+      content="default-src 'self'; script-src 'self'"
+    />
+    <title>Run Python Script</title>
+  </head>
+  <body>
+    <h1>Electron Python Shell Example</h1>
+    <button id="run-python">Run Python Script</button>
+    <p id="output">Output will appear here after script execution.</p>
+    <script src="./renderer.js"></script>
+  </body>
 </html>
