--- frontend/my-electron-app/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="Content-Security-Policy" content="default-src 'self'; script-src 'self'">
-    <title>Run Python Script</title>
-</head>
-<body>
-    <h1>Electron Python Shell Example</h1>
-    <button id="run-python">Run Python Script</button>
-    <p id="output">Output will appear here after script execution.</p>
-    <script src="./renderer.js"></script>
-</body>
-</html>
