--- frontend/my-electron-app/main.js
@@ -1,49 +1,0 @@
-const { app, BrowserWindow, ipcMain } = require('electron');
-const { exec } = require('child_process');
-const path = require('path');
-
-let mainWindow;
-
-function createWindow() {
-    // Create the browser window.
-    mainWindow = new BrowserWindow({
-        width: 800,
-        height: 600,
-        webPreferences: {
-            preload: path.join(__dirname, 'preload.js')
-        }
-    });
-
-    // and load the index.html of the app.
-    mainWindow.loadFile('index.html');
-}
-
-app.whenReady().then(createWindow);
-
-app.on('window-all-closed', () => {
-    if (process.platform !== 'darwin') {
-        app.quit();
-    }
-});
-
-app.on('activate', () => {
-    if (BrowserWindow.getAllWindows().length === 0) {
-        createWindow();
-    }
-});
-
-ipcMain.on('run-python-script', (event) => {
-    exec('python script.py', (error, stdout, stderr) => {
-        if (error) {
-            console.error(`exec error: ${error}`);
-            event.reply('python-script-response', `Error: ${error.message}`);
-            return;
-        }
-        if (stderr) {
-            console.error(`stderr: ${stderr}`);
-            event.reply('python-script-response', `Error: ${stderr}`);
-            return;
-        }
-        event.reply('python-script-response', stdout);
-    });
-});
