--- frontend/my-electron-app/renderer.js
@@ -1,12 +1,0 @@
-document.addEventListener('DOMContentLoaded', () => {
-  const runPythonBtn = document.getElementById('run-python');
-  const outputArea = document.getElementById('output');
-
-  runPythonBtn.addEventListener('click', () => {
-      window.electronAPI.runPythonScript();
-  });
-
-  window.electronAPI.handlePythonScriptResponse((message) => {
-      outputArea.textContent = message;
-  });
-});
