--- frontend/my-electron-app/script.py
@@ -1,1 +1,0 @@
-print("Have fun!!!")
