+++ frontend/script.py
@@ -1,0 +1,1 @@
+print("Have fun!!!")
